+++ src/react/shared/utils.js
@@ -133,6 +133,11 @@ export function getExtraAttrs(props = {}) {
 let routerIdCounter = 0;
 let routerComponentIdCounter = 0;
 
+export function unsetRouterIds() {
+  routerIdCounter = 0;
+  routerComponentIdCounter = 0;
+}
+
 export function getRouterId() {
   routerIdCounter += 1;
   return `${now()}_${routerIdCounter}`;