+++ src/svelte/shared/utils.js
@@ -91,6 +91,11 @@ export function createEmitter(createEventDispatcher, props) {
 let routerIdCounter = 0;
 let routerComponentIdCounter = 0;
 
+export function unsetRouterIds() {
+  routerIdCounter = 0;
+  routerComponentIdCounter = 0;
+}
+
 export function getRouterId() {
   routerIdCounter += 1;
   return `${now()}_${routerIdCounter}`;